415: Migration from V12 to V13 is slow and resource-heavy
