--- backend/core/src/main/kotlin/com/ritense/valtimo/domain/ValtimoApplicationProperty.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright 2015-2024 Ritense BV, the Netherlands.
- *
- * Licensed under EUPL, Version 1.2 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" basis,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package com.ritense.valtimo.domain
-
-import jakarta.persistence.Column
-import jakarta.persistence.Entity
-import jakarta.persistence.Id
-import jakarta.persistence.Table
-
-@Entity
-@Table(name = "valtimo_application_property")
-data class ValtimoApplicationProperty(
-    @Id
-    @Column(name = "property_key", nullable = false, updatable = false)
-    val propertyKey: String,
-    @Column(name = "property_value", nullable = false, updatable = false)
-    val propertyValue: String
-)
